requirements.txt
Django
djangorestframework
drf-yasg
django-cors-headers
Pillow
django-nested-admin